backend/requirements.txt
bidict==0.23.0
blinker==1.6.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
colorist==1.1.2
dnspython==2.4.2
eventlet==0.35.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.2.1
waitress==2.1.2
Werkzeug==2.3.6
wsproto==1.2.0
zipp==3.16.2
